+++ package.json
@@ -15,7 +15,7 @@
   "author": "Haeri",
   "license": "MIT",
   "devDependencies": {
-    "electron": "^11.2.3",
+    "electron": "^17.2.0",
     "electron-packager": "^15.3.0",
     "eslint": "^7.2.0",
     "eslint-config-airbnb-base": "^14.2.1",
